Description
For this file, I imported a list of client locations. I then obtained their respective calling codes by using the geography function in the data tab. Then, I clicked the drop down arrow next to the country name and selected "calling code". I pasted the format from the first cell in the country column to the rest of the cells in the same column to obtain all the calling codes. I then saved the workbook as a .csv file to make importing it into my algorithm easier.