Client Location Base Data
Originally, I started with just a list of country names. In order to get their respective calling codes, I utilized the geography function in the "data" tab in Excel, and pasted that formatting for every country using specific Excel shortcuts. From implementing the geography function, a drop-down arrow appears next to the country name. From there, a list of options is given, including the calling code. I pasted this formating for all the countries once again to get all the calling codes.