How to convert from calling code to country name
This file contains the sorted calling codes. Now, to get these calling codes back to their respective country, you have to utilize the file "All Country Calling Code Data". Theoretically, both that file and this file should be in the same Excel workbook, but since the only way to show an Excel file on GitHub is by exporting it as a .csv, and the only way to export a .csv file is to have one sheet in your workbook, these are two separate files in this document. In order to convert, you would have to utilize the VLookup function in Excel.